+++ CHANGELOG.md
@@ -47,6 +47,8 @@ Changelog
 * Prefixed *getHardwareAddress*, *getInetAddresses*, *getInterfaceAddresses* (network) with ([NetworkInterface](http://developer.android.com/reference/java/net/NetworkInterface.html))
 * Prefixed *getByInetAddress*, *getByName*, *getNetworkInterfaces* (internet) with ([NetworkInterface](http://developer.android.com/reference/java/net/NetworkInterface.html))
 * Added restriction [NetworkInterface.getByIndex](http://developer.android.com/reference/java/net/NetworkInterface.html#getByIndex(int))
+* Prefixed *getDetailedState*, *getState*, *isConnected*, *isConnectedOrConnecting* with [NetworkInfo](http://developer.android.com/reference/android/net/NetworkInfo.html)
+* Moved *getExtraInfo* to *network*
 * Updated in application documentation
 * Updated Dutch translation
 * Updated German translation